Cats don't like theremin noises
A sound is now produced, mechanics are a bit better. Doesn't bend the pitch properly because I can't do the
maths in my head, or at all, it seems. Need to end up with a range so that Y-1 = 0, Y0 = 8192, and Y1 = 16383.
Also need to utilise the right hand and X axis better.